:arrow_up: Updated: Next.js 12.1.6 -> 12.2.0
