+++ package.json
@@ -58,7 +58,7 @@
     "i18next-browser-languagedetector": "^6.1.5",
     "i18next-http-backend": "^1.4.1",
     "js-file-download": "^0.4.12",
-    "next": "12.1.6",
+    "next": "12.2.0",
     "next-i18next": "^11.3.0",
     "prism-react-renderer": "^1.3.5",
     "react": "^18.2.0",
